Overlay: Add config_intrusiveBatteryLed
Looks like that NotificationsSettings checks wheter config_intrusiveBatteryLed
is enabled or that the user isn't the user owner.

My guess is that both situations return false and by so, NotificationsSettings
tries to find a preference that just isn't available (because flo doesn't
support battery LED) and returns an NPE.

Let's change that by adding at least one of the required options of the check.